[Common] CFile::Read():  Use standard memory size input type.

+++ Source/Common/File Class.h
@@ -40,7 +40,7 @@ public:
     virtual bool SetLength(uint32_t dwNewLen) = 0;
     virtual uint32_t GetLength() const = 0;
 
-    virtual uint32_t Read(void* lpBuf, uint32_t nCount) = 0;
+    virtual size_t Read(void* lpBuf, size_t nCount) = 0;
     virtual bool Write(const void* lpBuf, size_t nCount) = 0;
 
     virtual bool Flush() = 0;
@@ -78,7 +78,7 @@ public:
     virtual bool SetLength(uint32_t dwNewLen);
     virtual uint32_t GetLength() const;
 
-    virtual uint32_t Read(void* lpBuf, uint32_t nCount);
+    virtual size_t Read(void* lpBuf, size_t nCount);
     virtual bool Write(const void* lpBuf, size_t nCount);
 
     virtual bool Flush();
